test(storage): extract keys/peerstore inline tests to sibling file
Move peerstore key encoding unit tests into peerstore_tests.rs.
Preserve should_panic attributes, module path keys::peerstore::tests,
and exact string assertions.
